+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [0.3.0] - 2023-02-24
+
+### Changed
+- Renamed `kernel_sidecar.kernel.SidecarKernelClient` to `kernel_sidecar.client.KernelSidecarClient` for better accuracy / consistency
 
 ## [0.2.1] - 2023-02-24
+
 ### Added
 - `SidecarKernelClient.running_action` property to return what action the Kernel is probably handling. Useful for testing
 - Callback hook in the `SidecarKernelClient` to take action on ZMQ channel disconnect
@@ -16,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Significant refactors after working on integrating `kernel-sidecar` into a production app
+  - `Kernel` renamed to `SidecarKernelClient`
   - `request` models (`execute_request`, `kernel_info_request`, etc) are split out into their own module from `actions`
   - `actions` are now primarily responsible for knowing when they're complete (`awaitable`) and attaching 0-n `handlers`
   - `action.observe` removed since you can attach more than one `handler`. It was unclear when to use `observers` vs `handlers`